Fix console warning; move ConfirmDeleteButton file
